Major restructure performed
- Created State files

- Create Contexts files

- Added space constants

- Added action points constants

- Used InitSpace derive macros

- Created helpers file
