+++ programs/rpg/src/constants.rs
@@ -7,3 +7,10 @@ pub const IS_STUNNED_FLAG: u8 = 1 << 5;
 pub const IS_SLOWED_FLAG: u8 = 1 << 6;
 pub const IS_BLEEDING_FLAG: u8 = 1 << 7;
 pub const NO_EFFECT_FLAG: u8 = 0b00000000;
+pub const SPAWN_MONSTER_ACTION_POINTS: u64 = 5;
+pub const CREATE_PLAYER_ACTION_POINTS: u64 = 100;
+pub const ATTACK_ACTION_POINTS: u64 = 1;
+
+pub const MAX_INVENTORY_ITEMS: usize = 8;
+
+pub const ANCHOR_DISCRIMINATOR: usize = 8;
