+++ programs/rpg/src/contexts/mod.rs
@@ -1,0 +1,11 @@
+pub mod attack_monster;
+pub mod collect_points;
+pub mod create_game;
+pub mod create_player;
+pub mod spawn_monster;
+
+pub use attack_monster::*;
+pub use collect_points::*;
+pub use create_game::*;
+pub use create_player::*;
+pub use spawn_monster::*;
