+++ programs/rpg/src/helpers.rs
@@ -1,0 +1,27 @@
+use anchor_lang::{prelude::*, system_program::{transfer, Transfer}};
+
+use crate::Player;
+
+pub fn spend_action_points<'info>(
+    action_points: u64, 
+    player_account: &mut Account<'info, Player>,
+    player: &AccountInfo<'info>, 
+    system_program: &AccountInfo<'info>, 
+) -> Result<()> {
+
+    player_account.action_points_spent = player_account.action_points_spent.checked_add(action_points).unwrap();
+    player_account.action_points_to_be_collected = player_account.action_points_to_be_collected.checked_add(action_points).unwrap();
+
+    let cpi_context = CpiContext::new(
+        system_program.clone(), 
+        Transfer {
+            from: player.clone(),
+            to: player_account.to_account_info().clone(),
+        }
+    );
+    transfer(cpi_context, action_points)?;
+
+    msg!("Minus {} action points", action_points);
+
+    Ok(())
+}
